Refactored structure of the project.
